[refactor] Altera o o funcionanmento da página suppplier e adiciona novo modal ao suppliers
